Add a json file to help map state from input to output
In some cases we need to return data that was previously mocked.
For example, if you create a bucket, or run an instance, then describe
that created resource you would expect the same data to be returned.

That's not terribly difficult but building some way to map that data
in a simple to edit format so that it can be customized per service
is necessary. Some data cross cut concerns are complex and need to
be modeled.